Tight permission-value check
Previously we accepted any value for permission-value parameter and
assumed it contained a valid value. But if it didn't, we tried to
store a NULL value in the database, which produced an error (the
column for that value is defined as NOT NULL).

Now we have a pre-condition check, and refuse any non-valid value.