Fix limitation of rbac_role_assigment primary key
Currently the primary key for rbac_role_assignment is defined as
follows in rbac-tables.pg.up.sql:

    PRIMARY KEY (role_id, context_id)

That definition prevents a role from begin assigned to more than one
user on a given context. Which is an unexpected and undesired
limitation. In order to overcome that limitation we need to add the
user user_id column to the primary key too.

    PRIMARY KEY (role_id, context_id, user_id)

[Re: #1]