Avoid the need to recover UUIDs from PG Json Obj.
[Re #3]

As we were building json objects from the function to get user's
role-assignments in a given context, we were losing the ID types of
`rbac_role` and `rbac_context` tables, so we were parsing the data from
string format to UUID one. This, prevents us to make more flexible the
definition of id types for both `user-id` and `resource-id` properties, that
might be different, as it is app-domain-specific.

Now, we have refactored the code to avoid using those JSON Object
build-related transformations, so we build the same expected response
structure with the code, in a way don't assume these grouped properties are
the only existing ones for each role-assigment record.

Besides, as we don't need parsing trasnformations anymore, we can make the
code independent of PostgreSQL (we still have some dependencies because of
the way we make tests, but we could fix that in the future).

Added new test to cover the target method refactored.