Support extra opts for resourc. and user id types
[Re #3]

Before we were only supporting UUID as type for Resource and User id props,
but now we are supporting also string and integer formats, since the usage of
this library can imply different app-domain-specific schemas for these
entities.

We do not care about the id types for the rest of the entities, as they are
not connected to any app-specific domain and we prefer to work with UUIDs
explictly in those cases.